Move well includes to subdirectory Well/
