+++ opm/core/wells/WellsGroup.hpp
@@ -25,7 +25,7 @@
 #include <opm/grid/UnstructuredGrid.h>
 #include <opm/core/props/BlackoilPhases.hpp>
 
-#include <opm/parser/eclipse/EclipseState/Schedule/Well.hpp>
+#include <opm/parser/eclipse/EclipseState/Schedule/Well/Well.hpp>
 #include <opm/parser/eclipse/EclipseState/Schedule/Group.hpp>
 
 #include <string>
